chore: organize files
- add module name consistency
